Split block-kit tests into multiple files
